hgw_frontend/requirements.txt
Django==2.2.5
mock==2.0.0
djangorestframework==3.10.3
django-oauth-toolkit==0.12.0
django-cors-middleware==1.3.1
requests==2.22.0
requests_oauthlib==0.8.0
kafka_python==1.4.6
pycryptodomex==3.4.7
coreapi==2.3.3
PyYAML==5.1.1
djangosaml2==0.17.2
psycopg2-binary